Change Blockweight data type
size_t -> uint64_t

+++ src/blockchain_db/blockchain_db.h
@@ -364,7 +364,7 @@ private:
    * @param blk_hash the hash of the block
    */
   virtual void add_block( const block& blk
-                , size_t block_weight
+                , uint64_t block_weight
                 , uint64_t long_term_block_weight
                 , const difficulty_type& cumulative_difficulty
                 , const uint64_t& coins_generated
@@ -798,7 +798,7 @@ public:
    * @return the height of the chain post-addition
    */
   virtual uint64_t add_block( const block& blk
-                            , size_t block_weight
+                            , uint64_t block_weight
                             , uint64_t long_term_block_weight
                             , const difficulty_type& cumulative_difficulty
                             , const uint64_t& coins_generated
